pratice.py worked in the get_training_data py
