+++ customerclustering/practice.py
@@ -4,7 +4,7 @@ from db_connection import Db
 
 
 class Practice:
-    def __init__(self, conn,rows):
+    def __init__(self, conn):
         # Import data only once
 
         self.conn=conn
